Refactor
All Components and their corresponding parts (subcomponents) are now self contained in their own folders.
Seperation of concerns is the main principle behind this structure.
Each Component folder will contain an index file which handles the grouping of it's multiple parts, consisting in: Type-definitions, subcomponents, and possibly styles & unit tests, among other parts.
