+++ src/components/AnimatedStyleUpdateExample/index.jsx
