+++ src/components/Trimmer/index.tsx
@@ -1,0 +1,12 @@
+import { StyleSheet, Text, View } from 'react-native';
+import React from 'react';
+
+export default function index() {
+  return (
+    <View>
+      <Text></Text>
+    </View>
+  );
+}
+
+const styles = StyleSheet.create({});
