+++ src/components/index.js
@@ -3,3 +3,4 @@ export { default as Reactive1 } from './Reactive1'
 export { default as Reactive2 } from './Reactive2'
 export { default as Reactive3 } from './Reactive3'
 export { default as Reactive7 } from './Reactive7'
+export { default as Trimmer } from './Trimmer'
